# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CMBCluster API",
    "description": "Multi-tenant Streamlit Platform for Cosmology Research",
    "version": "1.0.0"
  },
  "paths": {
...
    "/storage/list": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "List User Storages",
        "description": "List all storage buckets for the current user",
        "operationId": "list_user_storages_storage_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/storage/create": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "Create Storage",
        "description": "Create a new storage bucket for the user",
        "operationId": "create_storage_storage_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageCreationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/storage/{storage_id}": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Get Storage Details",
        "description": "Get detailed information about a specific storage bucket",
        "operationId": "get_storage_details_storage__storage_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "storage"
        ],
        "summary": "Delete Storage",
        "description": "Delete a storage bucket",
        "operationId": "delete_storage_storage__storage_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BucketOperationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/storage/{storage_id}/upload": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "Upload File To Storage",
        "description": "Upload a file to storage bucket",
        "operationId": "upload_file_to_storage_storage__storage_id__upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_to_storage_storage__storage_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/storage/{storage_id}/download/{object_path}": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Download File From Storage",
        "description": "Download a file from storage bucket",
        "operationId": "download_file_from_storage_storage__storage_id__download__object_path__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          },
          {
            "name": "object_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/storage/{storage_id}/list": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "List Storage Objects",
        "description": "List objects in storage bucket",
        "operationId": "list_storage_objects_storage__storage_id__list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prefix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/storage/{storage_id}/objects/{object_path}": {
      "delete": {
        "tags": [
          "storage"
        ],
        "summary": "Delete Storage Object",
        "description": "Delete an object from storage bucket",
        "operationId": "delete_storage_object_storage__storage_id__objects__object_path__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          },
          {
            "name": "object_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_to_storage_storage__storage_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_to_storage_storage__storage_id__upload_post"
      },
      "BucketOperationRequest": {
        "properties": {
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "backup_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup Location"
          }
        },
        "type": "object",
        "required": [
          "operation"
        ],
        "title": "BucketOperationRequest",
        "description": "Request model for bucket operations"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StorageCreationResponse": {
        "properties": {
          "storage": {
            "$ref": "#/components/schemas/UserStorage"
          },
          "bucket_metadata": {
            "type": "object",
            "title": "Bucket Metadata"
          },
          "permissions_configured": {
            "type": "boolean",
            "title": "Permissions Configured"
          }
        },
        "type": "object",
        "required": [
          "storage",
          "bucket_metadata",
          "permissions_configured"
        ],
        "title": "StorageCreationResponse",
        "description": "Response model for storage creation"
      },
      "StorageListResponse": {
        "properties": {
          "storages": {
            "items": {
              "$ref": "#/components/schemas/UserStorage"
            },
            "type": "array",
            "title": "Storages"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "usage_stats": {
            "$ref": "#/components/schemas/StorageUsageStats"
          }
        },
        "type": "object",
        "required": [
          "storages",
          "total_count",
          "usage_stats"
        ],
        "title": "StorageListResponse",
        "description": "Response model for storage listing"
      },
      "StorageMetadata": {
        "properties": {
          "storage": {
            "$ref": "#/components/schemas/UserStorage"
          },
          "bucket_metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket Metadata"
          },
          "recent_activity": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Recent Activity"
          },
          "permissions_valid": {
            "type": "boolean",
            "title": "Permissions Valid",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "storage"
        ],
        "title": "StorageMetadata",
        "description": "Detailed storage metadata for management"
      },
      "StorageRequest": {
        "properties": {
          "storage_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StorageType"
              }
            ],
            "default": "cloud_storage"
          },
          "storage_class": {
            "type": "string",
            "title": "Storage Class",
            "default": "STANDARD"
          },
          "custom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "title": "StorageRequest",
        "description": "Request model for creating storage"
      },
      "StorageStatus": {
        "type": "string",
        "enum": [
          "creating",
          "active",
          "deleting",
          "failed",
          "unknown"
        ],
        "title": "StorageStatus",
        "description": "Storage status states"
      },
      "StorageType": {
        "type": "string",
        "enum": [
          "pvc",
          "cloud_storage",
          "fuse"
        ],
        "title": "StorageType",
        "description": "Storage backend types"
      },
      "StorageUsageStats": {
        "properties": {
          "total_buckets": {
            "type": "integer",
            "title": "Total Buckets"
          },
          "total_size_bytes": {
            "type": "integer",
            "title": "Total Size Bytes"
          },
          "active_buckets": {
            "type": "integer",
            "title": "Active Buckets"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "total_buckets",
          "total_size_bytes",
          "active_buckets",
          "last_updated"
        ],
        "title": "StorageUsageStats",
        "description": "Storage usage statistics"
      },
...
      "UserStorage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "bucket_name": {
            "type": "string",
            "title": "Bucket Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "storage_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StorageType"
              }
            ],
            "default": "cloud_storage"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StorageStatus"
              }
            ],
            "default": "creating"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_accessed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Accessed"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "default": 0
          },
          "object_count": {
            "type": "integer",
            "title": "Object Count",
            "default": 0
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "storage_class": {
            "type": "string",
            "title": "Storage Class",
            "default": "STANDARD"
          },
          "versioning_enabled": {
            "type": "boolean",
            "title": "Versioning Enabled",
            "default": true
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "bucket_name",
          "display_name",
          "created_at",
          "location"
        ],
        "title": "UserStorage",
        "description": "User storage bucket metadata"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}